support for active_record 3.1 and 3.2
