+++ .travis.yml
@@ -2,6 +2,15 @@ script: "bundle exec rake test"
 rvm:
   - 1.8.7
   - 1.9.3
+  - 2.0
 gemfile:
-  - gemfiles/activerecord-2.3.gemfile
-  - gemfiles/activerecord-3.0.gemfile
+  - gemfiles/rails2.3.gemfile
+  - gemfiles/rails3.0.gemfile
+  - gemfiles/rails3.1.gemfile
+  - gemfiles/rails3.2.gemfile
+matrix:
+  exclude:
+    - rvm: 2.0
+      gemfile: gemfiles/rails2.3.gemfile
+    - rvm: 2.0
+      gemfile: gemfiles/rails3.0.gemfile
