+++ Appraisals
@@ -1,7 +1,15 @@
 appraise "rails2.3" do
-  gem "activerecord", "2.3.14", :require => 'active_record'
+  gem "activerecord", "~> 2.3.14", :require => 'active_record'
 end
 
 appraise "rails3.0" do
-  gem "activerecord", "3.0.15", :require => 'active_record'
+  gem "activerecord", "~> 3.0.15", :require => 'active_record'
+end
+
+appraise "rails3.1" do
+  gem "activerecord", "~> 3.1.0", :require => 'active_record'
+end
+
+appraise "rails3.2" do
+  gem "activerecord", "~> 3.2.0", :require => 'active_record'
 end
