+++ gemfiles/rails2.3.gemfile.lock
@@ -1,53 +1,53 @@
 PATH
-  remote: /Users/mgrosser/code/tools/active_record_inherit_assoc
+  remote: /Users/staugaard/code/active_record_inherit_assoc
   specs:
-    active_record_inherit_assoc (0.0.6)
+    active_record_inherit_assoc (1.0.0)
       activerecord
 
 GEM
   remote: http://rubygems.org/
   specs:
-    actionmailer (2.3.14)
-      actionpack (= 2.3.14)
-    actionpack (2.3.14)
-      activesupport (= 2.3.14)
+    actionmailer (2.3.18)
+      actionpack (= 2.3.18)
+    actionpack (2.3.18)
+      activesupport (= 2.3.18)
       rack (~> 1.1.0)
-    activerecord (2.3.14)
-      activesupport (= 2.3.14)
-    activeresource (2.3.14)
-      activesupport (= 2.3.14)
-    activesupport (2.3.14)
-    appraisal (0.4.1)
+    activerecord (2.3.18)
+      activesupport (= 2.3.18)
+    activeresource (2.3.18)
+      activesupport (= 2.3.18)
+    activesupport (2.3.18)
+    appraisal (0.5.2)
       bundler
       rake
+    bump (0.4.1)
     columnize (0.3.6)
-    debugger (1.1.4)
+    debugger (1.5.0)
       columnize (>= 0.3.1)
-      debugger-linecache (~> 1.1.1)
-      debugger-ruby_core_source (~> 1.1.3)
-    debugger-linecache (1.1.2)
-      debugger-ruby_core_source (>= 1.1.1)
-    debugger-ruby_core_source (1.1.3)
+      debugger-linecache (~> 1.2.0)
+      debugger-ruby_core_source (~> 1.2.0)
+    debugger-linecache (1.2.0)
+    debugger-ruby_core_source (1.2.0)
     git (1.2.5)
     jeweler (1.8.4)
       bundler (~> 1.0)
       git (>= 1.2.5)
       rake
       rdoc
-    json (1.7.3)
+    json (1.7.7)
     linecache (0.46)
       rbx-require-relative (> 0.0.4)
-    rack (1.1.3)
-    rails (2.3.14)
-      actionmailer (= 2.3.14)
-      actionpack (= 2.3.14)
-      activerecord (= 2.3.14)
-      activeresource (= 2.3.14)
-      activesupport (= 2.3.14)
+    rack (1.1.6)
+    rails (2.3.18)
+      actionmailer (= 2.3.18)
+      actionpack (= 2.3.18)
+      activerecord (= 2.3.18)
+      activeresource (= 2.3.18)
+      activesupport (= 2.3.18)
       rake (>= 0.8.3)
-    rake (0.9.2.2)
+    rake (10.0.4)
     rbx-require-relative (0.0.9)
-    rdoc (3.12)
+    rdoc (4.0.1)
       json (~> 1.4)
     ruby-debug (0.10.4)
       columnize (>= 0.1)
@@ -57,18 +57,19 @@ GEM
     shoulda (3.0.1)
       shoulda-context (~> 1.0.0)
       shoulda-matchers (~> 1.0.0)
-    shoulda-context (1.0.0)
+    shoulda-context (1.0.2)
     shoulda-matchers (1.0.0)
-    sqlite3 (1.3.6)
-    test-unit (2.5.1)
+    sqlite3 (1.3.7)
+    test-unit (2.5.4)
 
 PLATFORMS
   ruby
 
 DEPENDENCIES
   active_record_inherit_assoc!
-  activerecord (= 2.3.14)
+  activerecord (~> 2.3.14)
   appraisal
+  bump
   bundler
   debugger
   jeweler
