+++ gemfiles/rails3.2.gemfile
@@ -13,6 +13,6 @@ gem "jeweler"
 gem "ruby-debug", :platforms=>:ruby_18
 gem "debugger", :platforms=>:ruby_19
 gem "test-unit", ">=2.5.1"
-gem "activerecord", "~> 2.3.14", :require=>"active_record"
+gem "activerecord", "~> 3.2.0", :require=>"active_record"
 
 gemspec :path=>"../"
