+++ gemfiles/rails3.2.gemfile.lock
@@ -7,37 +7,37 @@ PATH
 GEM
   remote: http://rubygems.org/
   specs:
-    abstract (1.0.0)
-    actionmailer (3.0.20)
-      actionpack (= 3.0.20)
-      mail (~> 2.2.19)
-    actionpack (3.0.20)
-      activemodel (= 3.0.20)
-      activesupport (= 3.0.20)
-      builder (~> 2.1.2)
-      erubis (~> 2.6.6)
-      i18n (~> 0.5.0)
-      rack (~> 1.2.5)
-      rack-mount (~> 0.6.14)
-      rack-test (~> 0.5.7)
-      tzinfo (~> 0.3.23)
-    activemodel (3.0.20)
-      activesupport (= 3.0.20)
-      builder (~> 2.1.2)
-      i18n (~> 0.5.0)
-    activerecord (3.0.20)
-      activemodel (= 3.0.20)
-      activesupport (= 3.0.20)
-      arel (~> 2.0.10)
-      tzinfo (~> 0.3.23)
-    activesupport (3.0.20)
+    actionmailer (3.2.13)
+      actionpack (= 3.2.13)
+      mail (~> 2.5.3)
+    actionpack (3.2.13)
+      activemodel (= 3.2.13)
+      activesupport (= 3.2.13)
+      builder (~> 3.0.0)
+      erubis (~> 2.7.0)
+      journey (~> 1.0.4)
+      rack (~> 1.4.5)
+      rack-cache (~> 1.2)
+      rack-test (~> 0.6.1)
+      sprockets (~> 2.2.1)
+    activemodel (3.2.13)
+      activesupport (= 3.2.13)
+      builder (~> 3.0.0)
+    activerecord (3.2.13)
+      activemodel (= 3.2.13)
+      activesupport (= 3.2.13)
+      arel (~> 3.0.2)
+      tzinfo (~> 0.3.29)
+    activesupport (3.2.13)
+      i18n (= 0.6.1)
+      multi_json (~> 1.0)
     appraisal (0.5.2)
       bundler
       rake
-    arel (2.0.10)
+    arel (3.0.2)
     bourne (1.4.0)
       mocha (~> 0.13.2)
-    builder (2.1.2)
+    builder (3.0.4)
     bump (0.4.1)
     columnize (0.3.6)
     debugger (1.5.0)
@@ -46,20 +46,20 @@ GEM
       debugger-ruby_core_source (~> 1.2.0)
     debugger-linecache (1.2.0)
     debugger-ruby_core_source (1.2.0)
-    erubis (2.6.6)
-      abstract (>= 1.0.0)
+    erubis (2.7.0)
     git (1.2.5)
-    i18n (0.5.0)
+    hike (1.2.2)
+    i18n (0.6.1)
     jeweler (1.8.4)
       bundler (~> 1.0)
       git (>= 1.2.5)
       rake
       rdoc
+    journey (1.0.4)
     json (1.7.7)
     linecache (0.46)
       rbx-require-relative (> 0.0.4)
-    mail (2.2.19)
-      activesupport (>= 2.3.6)
+    mail (2.5.3)
       i18n (>= 0.4.0)
       mime-types (~> 1.16)
       treetop (~> 1.4.8)
@@ -67,11 +67,12 @@ GEM
     mime-types (1.23)
     mocha (0.13.3)
       metaclass (~> 0.0.1)
+    multi_json (1.7.2)
     polyglot (0.3.3)
-    rack (1.2.8)
-    rack-mount (0.6.14)
-      rack (>= 1.0.0)
-    rack-test (0.5.7)
+    rack (1.4.5)
+    rack-cache (1.2)
+      rack (>= 0.4)
+    rack-test (0.6.2)
       rack (>= 1.0)
     rails (0.9.5)
       actionmailer (>= 0.6.1)
@@ -94,8 +95,14 @@ GEM
     shoulda-matchers (1.5.6)
       activesupport (>= 3.0.0)
       bourne (~> 1.3)
+    sprockets (2.2.2)
+      hike (~> 1.2)
+      multi_json (~> 1.0)
+      rack (~> 1.0)
+      tilt (~> 1.1, != 1.3.0)
     sqlite3 (1.3.7)
     test-unit (2.5.4)
+    tilt (1.4.0)
     treetop (1.4.12)
       polyglot
       polyglot (>= 0.3.1)
@@ -106,7 +113,7 @@ PLATFORMS
 
 DEPENDENCIES
   active_record_inherit_assoc!
-  activerecord (~> 3.0.15)
+  activerecord (~> 3.2.0)
   appraisal
   bump
   bundler
